New option: --csv
With this option, three CSV files are generated to be loaded into a database. Method GenerateTestDataset has been modified in order to add more DICOM tags to the studies, series, and images. New function has been added to get a random value of the CT ImageType. Variables NumberOfStudyRelatedInstances and imageCount are hardcoded for some image types.